cleanups, changed style of facebook button, added sign up button
