+++ .idea/misc.xml
@@ -3,10 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="..\:/Users/HANIKO/.gradle/caches/transforms-3/fa52f6e5bda100d69615edee20b800b5/transformed/jetified-facebook-common-12.1.0/res/drawable/com_facebook_favicon_blue.xml" value="0.21822916666666667" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.20277777777777778" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/drawable/ic_FirstPageBackground_background.xml" value="0.22135416666666666" />
+        <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/drawable/ic_facebook_logo.xml" value="0.21822916666666667" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/drawable/ic_google_logo.xml" value="0.21822916666666667" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/drawable/ic_launcher_background.xml" value="0.20277777777777778" />
+        <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/drawable/ic_logo_login_with_email.xml" value="0.21458333333333332" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/mipmap-anydpi-v26/ic_FirstPageBackground.xml" value="0.22135416666666666" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/mipmap-anydpi-v26/ic_first_page_background.xml" value="0.20277777777777778" />
         <entry key="..\:/Users/HANIKO/AndroidStudioProjects/NoInstagram/app/src/main/res/mipmap-anydpi-v26/ic_first_page_background_round.xml" value="0.20277777777777778" />
