+++ app/src/main/java/com/example/noinstagram/ui/theme/Color.kt
@@ -7,3 +7,4 @@ val Purple500 = Color(0xFF6200EE)
 val Purple700 = Color(0xFF3700B3)
 val Teal200 = Color(0xFF03DAC5)
 val LightBlue = Color(0xFFC6D1E1)
+val FacebookColor = Color(0xFF3C5B9A)
