+++ app/src/main/res/drawable/ic_facebook_logo.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+    <path
+        android:fillColor="#3C5B9A"
+        android:pathData="M0,0h512v512h-512z" />
+    <path
+        android:fillColor="#fff"
+        android:pathData="M355.6,330l11.4,-74h-71v-48c0,-20.2 9.9,-40 41.7,-40H370v-63s-29.3,-5 -57.3,-5c-58.5,0 -96.7,35.4 -96.7,99.6V256h-65v74h65v182h80V330h59.6z" />
+</vector>
